+++ Assets/Tests/InputSystem/Plugins/PlayerInputTests.cs
@@ -5,7 +5,6 @@ using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.LowLevel;
-using UnityEngine.InputSystem.PlayerInput;
 using UnityEngine.InputSystem.UI;
 using UnityEngine.InputSystem.Users;
 using UnityEngine.InputSystem.Processors;
@@ -305,6 +304,27 @@ internal class PlayerInputTests : InputTestFixture
         playerInput2.actions = actions;
 
         Assert.That(playerInput1.actions, Is.Not.SameAs(playerInput2.actions));
+    }
+
+    [Test]
+    [Category("PlayerInput")]
+    public void PlayerInput_AssigningSameActionsToDifferentPlayers_DuplicatesOverrides()
+    {
+        var go1 = new GameObject();
+        var playerInput1 = go1.AddComponent<PlayerInput>();
+
+        var go2 = new GameObject();
+        var playerInput2 = go2.AddComponent<PlayerInput>();
+
+        var actions = InputActionAsset.FromJson(kActions);
+        actions.actionMaps[0].actions[0].ApplyBindingOverride(0, "<Gamepad>/buttonNorth");
+
+        playerInput1.actions = actions;
+        playerInput2.actions = actions;
+
+        Assert.That(playerInput1.actions, Is.Not.SameAs(playerInput2.actions));
+        Assert.That(playerInput1.actions.actionMaps[0].actions[0].bindings[0].overridePath, Is.SameAs("<Gamepad>/buttonNorth"));
+        Assert.That(playerInput2.actions.actionMaps[0].actions[0].bindings[0].overridePath, Is.SameAs("<Gamepad>/buttonNorth"));
     }
 
     [Test]