+++ Packages/com.unity.inputsystem/InputSystem/Actions/InputActionRebindingExtensions.cs
@@ -76,8 +76,8 @@ namespace UnityEngine.InputSystem
 
         public static void ApplyBindingOverride(this InputAction action, int bindingIndex, string path)
         {
-            if (string.IsNullOrEmpty(path))
-                throw new ArgumentException("Binding path cannot be null or empty", nameof(path));
+            if (path == null)
+                throw new ArgumentException("Binding path cannot be null", nameof(path));
             ApplyBindingOverride(action, bindingIndex, new InputBinding {overridePath = path});
         }
 
@@ -894,7 +894,7 @@ namespace UnityEngine.InputSystem
                 m_Flags &= ~Flags.OnAfterUpdateHooked;
             }
 
-            private void OnAfterUpdate(InputUpdateType updateType)
+            private void OnAfterUpdate()
             {
                 // If we don't have a match yet but we have a timeout and have expired it,
                 // cancel the operation.
@@ -932,7 +932,7 @@ namespace UnityEngine.InputSystem
                         // We have a callback. Give it a shot to generate a path. If it doesn't,
                         // fall back to our default logic.
                         var newPath = m_OnGeneratePath(selectedControl);
-                        if (!string.IsNullOrEmpty(path))
+                        if (!string.IsNullOrEmpty(newPath))
                             path = newPath;
                         else if ((m_Flags & Flags.DontGeneralizePathOfSelectedControl) == 0)
                             path = GeneratePathForControl(selectedControl);
@@ -1058,8 +1058,8 @@ namespace UnityEngine.InputSystem
             private Func<InputControl, string> m_OnGeneratePath;
             private Func<InputControl, InputEventPtr, float> m_OnComputeScore;
             private Action<RebindingOperation, string> m_OnApplyBinding;
-            private Action<InputUpdateType> m_OnAfterUpdateDelegate;
             private Action<InputEventPtr, InputDevice> m_OnEventDelegate;
+            private Action m_OnAfterUpdateDelegate;
             private InputControlLayout.Cache m_LayoutCache;
             private StringBuilder m_PathBuilder;
             private Flags m_Flags;