+++ Packages/com.unity.inputsystem/InputSystem/Plugins/PlayerInput/PlayerNotifications.cs
@@ -1,4 +1,4 @@
-namespace UnityEngine.InputSystem.PlayerInput
+namespace UnityEngine.InputSystem
 {
     /// <summary>
     /// Determines how the triggering of an action or other input-related events are relayed to other GameObjects.